Move retry logic to client module
